refactor: move useDarkMode to its own file
